Trait kyles_algorithm_crate::base::basic_dict::BasicDict

pub trait BasicDict<T>: Membership<T> {
    fn obtain(y: T) -> Result<T, ItemNotFoundError>;

    fn insert(x: T) -> Result<(), Error>;

    fn delete(x: T) -> Result<(), ItemNotFoundError>;
}

Represents a basic dictionary

Required Methods

fn obtain(y: T) -> Result<T, ItemNotFoundError>

Retrieve an item from the dictionary with membership_equals Returns a result with the item, or an ItemNotFoundError

Arguments

y - The item to retrieve

fn insert(x: T) -> Result<(), Error>

Insert an item into the dictionary Returns ContainerFullError if the container is full, or DuplicateItemsError if there is a duplicate

Arguments

x - The item to insert

fn delete(x: T) -> Result<(), ItemNotFoundError>

Delete the specified item Returns ItemNotFoundError if the item is not in the dict

Arguments

x - The item to delete