Trait kyles_algorithm_crate::base::dispenser::Dispenser

pub trait Dispenser<T>: Container + Cursor<T> {
    fn insert(x: T) -> Result<(), Error>;

    fn delete_item() -> Result<(), NoCurrentItemError>;
}

A container that keeps track of the current item as inserts are made Only the current item can be deleted

Required Methods

fn insert(x: T) -> Result<(), Error>

Insert an item into the structure Returns a DuplicateItemError or ContainerFullError

Arguments

x - The item to insert into the structure

fn delete_item() -> Result<(), NoCurrentItemError>

Delete the current item from the structure Returns a NoCurrentItemError if there is no current item