Struct kyles_algorithm_crate::base::ndpoint::NDPoint

pub struct NDPoint { /* fields omitted */ }

A representation of an N-dimensional point

Methods

impl NDPoint

Base functions for the NDPoint

pub fn new(dim: i32) -> Self

Create a new N-Dimensional point at the origin

Arguments

dim - The dimensionality of the point

Example

// creates a new 3-dimensional point at the origin
let nd_point = NDPoint::new(3);

pub fn from_coordinate(pt: Vec<f32>) -> Self

Create a new point from the given coordinate, given as a vector of floats. The dimensionality of the point is equal to the length of the vector

Arguments

pt - A vector of floats representing the coordinates for the point

Example

// creates a new 3-dimensional point at (1.5, 2.5, 3.5)
let nd_point = NDPoint::from_coordinate(vec![1.5, 2.5, 3.5]);

pub fn dim(&self) -> i32

Get the dimensionality of the point

pub fn set_point(&mut self, pt: Vec<f32>) -> Result<(), InvalidArgumentError>

Set the coordinates of this point Returns InvalidArgumentError if the dimension of the given point is less than 1

Arguments

pt - A vector of floats representing the coordinates to give this point

Example

// creates a new 3-dimensional point at the origin
let mut nd_point = NDPoint::new(3);
// Set the point to coordinates (3.2, 4.5, 6.7)
nd_point.set_point(vec![3.2, 4.5, 6.7]);

pub fn idx(&self, i: i32) -> Result<f32, InvalidArgumentError>

Return the ith coodinate of the point Returns InvalidArgumentError if the given index is greater than the dimensionality of the point

Arguments

i - The index of the coordinate to get

Example

// creates a new 3-dimensional point at the origin
let nd_point = NDPoint::from_coordinate(vec![5.0, 6.0, 7.0]);
let y = nd_point.idx(1).unwrap();
assert_eq!(y, 6.0);

pub fn compare_by_dim(
&self,
i: i32,
other: &NDPoint
) -> Result<i8, InvalidArgumentError>

Compares the ith coordinate of this point to another point Returns -1 if the ith coordinate of this point is smaller than that of 'other' Returns 0 if the ith coordinate of this point is equal to that of 'other' Returns 1 if the ith coordinate of this point is greater than that of 'other' Returns InvalidArgumentError if the other point has a different dimensionality or the given dimension is greater than that of this point

Arguments

i - The dimension of the points to compare

other - Another NDPoint to compare to

Example

use kyles_algorithm_crate::base::ndpoint::NDPoint;

// creates a new 3-dimensional point at the origin let nd_point = NDPoint::new(3); let nd_point2 = NDPoint::from_coordinate(vec![1.0, 1.0, 1.0]); // compare the points let res = nd_point.compare_by_dim(0, nd_point2); assert_eq(res, -1);