Struct kyles_algorithm_crate::graph::Graph

pub struct Graph<T> {
    pub nodes: HashMap<u32, Node<T>>,
}

Adjacency list implementation of a graph

Fields
nodes: HashMap<u32, Node<T>>
Methods

impl<T> Graph<T>

pub fn new() -> Self

Returns a new instance of a Graph

Example

// Creates a new graph
let graph: Graph<u32> = Graph::new();

pub fn add_node(&mut self, data: T) -> u32

Add a new node to the graph with supplied data. Returns the id of the node in the graph.

Arguments

data - The data to put in the node to be added

Example

// Creates a new graph
let mut graph = Graph::new();
 
// Adds a new node with value 10 to the graph
// node_id now contains the id to access that node in the graph
let node_id = graph.add_node(10);
 

pub fn add_directed_edge(&mut self, from: u32, to: u32)

Add a directed edge from one node to another

Arguments

from - The ID of the node the edge will start at to - The ID of the node the edge will go to

Example

// Create a graph
let mut graph = Graph::new();
 
// Add two nodes
let node1 = graph.add_node(1);
let node2 = graph.add_node(2);

// Add an edge from node1 to node 2
graph.add_directed_edge(node1, node2);

pub fn add_undirected_edge(&mut self, node1: u32, node2: u32)

Add an undirected edge between two nodes

Arguments

node1 - The ID of the first node to add an edge to node2 - The ID of the second node to add an edge to

Example

// Create a graph
let mut graph = Graph::new();
 
// Add two nodes
let node1 = graph.add_node(1);
let node2 = graph.add_node(2);

// Add an edge from node1 to node 2
graph.add_undirected_edge(node1, node2);

pub fn add_weighted_directed_edge(&mut self, from: u32, to: u32, weight: u32)

Add a weighted directed edge from one node to another

Arguments

from - The ID of the node the edge will come from to - The ID of the node the edge will go to weight - The weight of the edge

Example

// Create a graph
let mut graph = Graph::new();
 
// Add two nodes
let node1 = graph.add_node(1);
let node2 = graph.add_node(2);

// Add an edge of weight 10 from node1 to node 2
graph.add_weighted_directed_edge(node1, node2, 10);

pub fn add_weighted_undirected_edge(
&mut self,
node1: u32,
node2: u32,
weight: u32
)

Add a weighted undirected edge between two nodes

Arguments

node1 - The ID of the first node to add the edge to node2 - The ID of the second node to add the edge to weight - The weight for the edge

Example

// Create a graph
let mut graph = Graph::new();
 
// Add two nodes
let node1 = graph.add_node(1);
let node2 = graph.add_node(2);

// Add an edge of weight 10 from node1 to node 2
graph.add_weighted_undirected_edge(node1, node2, 10);

pub fn get_node(&self, node: u32) -> Option<&Node<T>>

Get a node from the graph

Arguments

'node' - The ID of the node to get from the graph

Example

// Create a graph
let mut graph = Graph::new();
 
// Add two nodes
let node1 = graph.add_node(1);
let node2 = graph.add_node(2);

// get node 1
let received_node = graph.get_node(node1);

pub fn get_node_mut(&mut self, node: u32) -> Option<&mut Node<T>>

Get a mutable node from the graph

Arguments

'node' - The ID of the node to get from the graph

Example

// Create a graph
let mut graph = Graph::new();
 
// Add two nodes
let node1 = graph.add_node(1);
let node2 = graph.add_node(2);

// get node 1
let received_node = graph.get_node_mut(node1);