Struct kyles_algorithm_crate::lists::arrayed_list_iterator::ArrayedListIterator

pub struct ArrayedListIterator<T> { /* fields omitted */ }

Methods

impl<T> ArrayedListIterator<T> where
T: Copy,

pub fn new(elements: Vec<T>, head: i32, tail: i32, num_el: u32) -> Self

Create a new Arrayed List iterator

Arguments

elements - The elements in the arrayed list to include in the iterator

head - The head of the list

tail - The tail of the list

num_el - The number of elements in the list

Example

 // Create a new ArrayedListIterator with the elements 1, 2, 3
 // The head of the list is the first element, the tail is the last, and there are
 // currently three elements in the list 
 let iter = ArrayedListIterator::new(vec![1, 2, 3], 0, 2, 3);
  

pub fn set_position(&mut self, pos: i32) -> Result<(), Box<Error>>

Set the current position of the iterator

Arguments

pos - The index of the position to set the iterator to

Example

 
// The current position of the iterator is head
let mut iter = ArrayedListIterator::new(vec![1, 2, 3], 0, 2, 3);
 
// Set the position to the second element, being 2 in this case
iter.set_position(1);

pub fn item(&self) -> Result<T, NoCurrentItemError>

Return the current item in the iterator

Example

let iter = ArrayedListIterator::new(vec![1, 2, 3], 0, 2, 3);

// Gets the current item in the list, 1 in this case
let item = iter.item();

pub fn item_exists(&self) -> bool

Check if there is a current item

Example

let iter = ArrayedListIterator::new(vec![1, 2, 3], 0, 2, 3);
if iter.item_exists(){
    println!("There is an item here!");
}