Struct kyles_algorithm_crate::tree::Tree

pub struct Tree<T> { /* fields omitted */ }

Representation of a tree structure

Methods

impl<T> Tree<T> where
T: Clone,

pub fn new(root_val: T) -> Self

Returns a new instance of a tree, with root_val as the value for the node

Arguments

root_val - The value to store at the root of this tree

Example

// Creates a new tree with the value 1 at its root
let tree = Tree::new(1);

pub fn add_subtree(&mut self, tree: Tree<T>)

Adds a given tree to the list of children of this tree

Arguments

tree - The tree to add to the list of children

Example

// Creates a new tree with the value 1 at its root
let mut tree = Tree::new(1);

let two = Tree::new(2);
let three = Tree::new(3);

// Add the two new trees to the root
tree.add_subtree(two);
tree.add_subtree(three);

// root.children now contains two and three

pub fn add_subtree_value(&mut self, value: T)

Add a new subtree to the root by specifying a value to store at the new child node

Arguments

value - The value to store at the new child node

Example

// Creates a new tree with the value 1 at its root
let mut tree = Tree::new(1);
 
tree.add_subtree_value(2);
tree.add_subtree_value(3);

// root.children now contains subtrees with values 2 and 3

pub fn postorder_traversal(&self) -> Vec<T>

Returns the values in the tree that are visited in a post-order traversal of the tree from the root

Arguments

&self - Borrows itself so that we can evaluate the child nodes and still be able to use the reference that called this function afterward

Example

// Usage of postorder traversal

// This will contain a vector of the values in the order they were visited
let postorder = root.postorder_traversal();
 

pub fn postorder_traversal_node(start_node: &Tree<T>) -> Vec<T>

Returns the values in the tree that are visited in a post-order traversal of the tree from the given node

Arguments

start_node - The sub tree to start the traversal at

pub fn preorder_traversal_node(start_node: &Tree<T>) -> Vec<T>

Returns the values in the tree that are visited in a pre-order traversal of the tree from the given node

Arguments

start_node - The sub tree to start the traversal at

pub fn preorder_traversal(&self) -> Vec<T>

Returns the values in the tree that are visited in a post-order traversal of the tree from the root

Arguments

&self - Borrows itself so that we can evaluate the child nodes and still be able to use the reference that called this function afterward

Example

// Usage of preorder traversal

// This will contain a vector of the values in the order they were visited
let postorder = root.preorder_traversal();